Added a missing dependency on libXvMC.
The xf86-video-intel needs libXvMC to be present when building with the
XVMC option enabled, This was missing and caused the port to not
compile the libI810XvMC* and libIntelXvMC* libraries even when the port
had the option enabled, causing a plist error.

This is now fixed, and the option is enabled by default again.