CI for the x11set branch.
The x11set branch is intended for testing and working on eventually providing the X11 binaries as sets for FreeBSD.

Split the CI into two tasks one for the master branch and one for x11set.

The x11set task's difference is in it setting the PREFIX directory of the build to /usr/X11R7, and putting its artifacts in a different directory.