Separated the X servers form flavors to 4 ports.
The Xnest, Xvfb, Xephyr and the X server were all crammed into the
xlibre-server port as 4 FLAVORs.
Separated them into 4 distinct ports for more clarity, ease of testing, and ...
 .

All of these ports now include the Makefile.common, a file that holds
common flags and configurations for all of them, or all of them except
the main server.

Makefile.version is used for setting the PORTVERSION across all 4 ports
and is included by Makefile.common.

xlibre-cat.mk now has a server category that sets the upstream of all 4
server ports as it did in the drivers category.
It can also set some flags that don't belong to Makefile.common for the
server category, but nothing is set there as of now.

The distinfo file and the folder for patches for all of the ports is set
to the xlibre-server's distinfo file and files/ folder respectively.