Added knobs for UDEV, the SUID wrapper & LIBUNWIND.
All these knobs are on by default.

The SUID knob is for disabling installation of the setuid wrapper around
the Xorg binary.

The UDEV knob is used for disabling all udev support, this makes XLibre
server not work with libinput and disables hotplug.

The LIBUNWIND knob is used for disabling use of libunwind for stack
traces in case of an error. This knob is shared across the other 3
serves.