Added support for DragonFlyBSD in CI.
The CI system now supports DragonFlyBSD too and will also create a pkg
repository for it.

This is done with a QEMU VM running on a Linux KVM container on Cirrus
CI. The VM is by the .ci/vm.sh script.

The DragonFlyBSD image is fetched from the Github
vmactions/dragonfly-builder repository. I plan to create a local image
builder, but I currently do not have much time for it.

The CI commands and instruction are moved from the main body of
.cirrus.yml file to .ci/run.sh, this script is ran on both FreeBSD and
DragonFlyBSD builds.

Few tiny C utilities are also written for the vm.sh script these are
also located in .ci directory.