Moved everything to the latest upstream version.
Both the servers and the drivers are moved to their latest version.

The XLibre upstream is now using an stable versioning and tagging system
for the drivers, so I changed all of the drivers PORTVERSIONs to the
latest release tag versions and removed the usage of GH_TAGNAME.

GH_TAGNAME is still used in a few drivers because their latest commits
were not yet tagged. In these drivers PORTREVISON is set to the latest
upstream tag version but their GH_TAGNAME is set to the latest commit
hash and PORTREVISION is set to one.