Added scfb & egalax drivers with XLibre flavors + More.
Added xf86-input-egalax and xf86-video-scfb 3rd party drivers with
XLibre flavors. These ports are not from XLibre, but they need to be
compiled against xlibre-server to be installable alongside it. So these
ports are now added and modified with two flavors available: xorg and
xlibre.

The xorg flavor for these ports acts just the same as the version of
them available in the main ports tree.

The xlibre flavor, which is the default for these ports in THIS
repository, will be built against xlibre-server and have a
dependency on it. The package name will also be prefixed with 'xlibre-';
So for example the xlibre flavor of xf86-video-scfb will generate a
package named xlibre-xf86-video-scfb that depends on xlibre-server
instead of xorg.

In addition, since all the ports built with GNU AutoTools, now respect
the proper server module directory location, a hackish section in
xlibre-cat.mk that set the module folder for AutoTool drivers has now
been removed.

Small fixed are also done in xlibre-cat.mk and xlibre-drivers to make
them respect this addition.

I'll hopefully will continue to add these 3rd party drivers into this
repository.