Added XLibre flavors for KDE ports.
KDE plasma 6 couldn't be installed alongside XLibre due to the dependence
of the port plasma6-plasma-desktop on X.Org drivers.

To solve this, I have added two flavors to these ports: 'xlibre' and 'xorg'.

The 'xorg' flavor acts as the main tree version of these ports and
depends on X.Org drivers.

The 'xlibre' flavor (which is the default in THIS tree) will depend on
the XLibre versions of the drivers and can be installed alongside XLibre
X server. The packages with this flavor will be prefixed with 'xlibre-'.

Sadly there is no support for DragonflyBSD yet, because the dports lack
essential dependencies for these KDE ports. This will hopefully get
fixed in the next release of dports in the following months.