Stopped setting the module directory for meson drivers.
A section in xlibre-cat.mk that set the module directory for drivers
that used the Meson build system has been removed.

It was really a ugly hack and was used just for the Intel video driver.

The Intel video driver has also been patched to obtain the module
directory form pkgconfig.