elographics: Updated to the latest version 25.0.1.
This driver has been updated to the latest release.

XLibre drivers now follow a semantic versioning system, and the major
version is synced with the XLibre server.

Driver CONFLICTS are now handled by the xlibre-cat framework.

The GNU_CONFIGURE_MANPREFIX line was unnecessary so removed it.

The driver path in the PLIST is now automatically chosen based on the
XLibre major server version.

(Accurate) portlint warnings have also been taken care of.