Added a separate port for the SCFB video driver.
This driver has been recently forked by the XLibre organization so we
can now create a new dedicated port for it, instead of adding FLAVORs to
the X.Org port.

This driver has been also updated to the latest release.

XLibre drivers now follow a semantic versioning system, and the major
version is synced with the XLibre server.

Driver CONFLICTS are now handled by the xlibre-cat framework.

The GNU_CONFIGURE_MANPREFIX line was unnecessary so removed it.

The driver path in the PLIST is now automatically chosen based on the
XLibre major server version.

(Accurate) portlint warnings have also been taken care of.