intel: Updated to the latest version 25.0.0.
This driver has been updated to the latest release.

XLibre drivers now follow a semantic versioning system, and the major
version is synced with the XLibre server.

Driver CONFLICTS are now handled by the xlibre-cat framework.

The driver path in the PLIST is now automatically chosen based on the
XLibre major server version.

Some portlint warnings have also been taken care of.

Added the patch file patch-rm-config.h that removes lines #including
config.h across all of the source files. The file has been removed in
the server sdk but the includes are still remaining, causing a build
error. This is not a standard patch file, it patches many files and
sometimes these files already have other patches, It was the only
clean solution.

The patch file patch-src_meson.build has been deleted because it was
merged into the upstream.