xorgxrdp*: Patched to work with XLibre 25.1.0.
Both xorgxrdp and xorgxrdp-devel drivers have been patched to work with
the new XLibre 25.1.0 release. The revision numbers of the XLibre
FLAVORs have been incremented to ensure updates.

The extra patch files extra-patch-module_rdpCapture.c and
 extra-patch-module_rdpEgl.c have been added to change the 'NullClient'
macro to 'NULL', this macro has been removed from the XLibre SDK because
it was always set to 'NULL'.