ci: Temporarily disable DragonFly Builds.
The DragonFly pkg is glitching out once again when trying to create a
repository, the actual builds have been completed in this step, but the
repo fails to build/hangs trying to.

The FreeBSD builds have been successful. This we let them hit the
artifact repositories and the DFBSD problem can be dealt with later.